requirements.txt
streamlit==1.27.0
boto3==1.28.63
python-dotenv==1.0.0
googletrans==3.1.0a0